`denote` modifications + introducing `hl-todo` and `cscope`
This patch introduces some utility Elisp functions I wrote to better
manage zettels within the `denote-directory`. There was also the need
to change keybindings for the introduction of the `cscope` client
for Emacs. Then, `hl-todo` serves as a convenient way to immediately
visualize key points to review in a piece of code/text.