Modifications to `exwm` to make it work on the work laptop
KDE started annoying me in a non-tolerable way. I was experiencing
some issues with EXWM on the work laptop for some Wayland-related crap
Debian automatically installed (which is even more annoying), which is
the reason why I decided not to use EXWM on it. However, things
changed: I fixed everything and now EXWM works flawlessly even on the
work laptop.