Moving to the new configuration structure
This is my new, entirely literate, Emacs configuration. It is based on
emacs-modules and custom-modules. The former are just source files
where I pull some packages of interest and add some tweaking over
them, while the latter are libraries where I write some custom code to
help me going in my day-by-day workflow with Emacs.

The main point of entry for source code is an `init.el` file where I
require all needed modules. The `mg-emacs.org` file is the one I
customize to generate the whole configuration, and serves as
documentation for it.