Complete refactoring of the `mg-bib.el` library
This commit introduces a substantial modification of the custom
library I use to manage my bibliographic entries. There are no new
functionalities, but just some refactoring for the code, that was
pretty ugly before. Now, the web pages handling is still to be
implemented, but it's not a huge amount of work, maybe one/two elisp
functions.